# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server to manage payments",
        "title": "Payment Simple API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/user": {
            "get": {
                "description": "Get lists of users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.FailedOperation"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a user with his name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create User",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "CreateUserDto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateUserDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.SucessfullOperation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.FailedOperation"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.FailedOperation"
                        }
                    }
                }
            }
        },
        "/user/{id}/loan": {
            "get": {
                "description": "Get lists user loans",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user loans",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.Loan"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.FailedOperation"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.FailedOperation"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.CreateUserDto": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.FailedOperation": {
            "type": "object",
            "properties": {
                "internalCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.Loan": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "debt": {
                    "type": "number"
                },
                "endDate": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "loanHistory": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.LoanHistory"
                    }
                },
                "quota": {
                    "type": "number"
                },
                "rate": {
                    "type": "number"
                },
                "startDate": {
                    "type": "string"
                },
                "targetName": {
                    "type": "string"
                },
                "targetSchemaId": {
                    "type": "string"
                },
                "term": {
                    "type": "integer"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "models.LoanHistory": {
            "type": "object",
            "properties": {
                "accumulated": {
                    "type": "number"
                },
                "monthDebt": {
                    "type": "number"
                },
                "monthEnd": {
                    "type": "string"
                },
                "monthStart": {
                    "type": "string"
                },
                "paymentId": {
                    "type": "string"
                }
            }
        },
...
        "models.SucessfullOperation": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}